refactored BedroomsDisplay to one file
